[fix] highscores 제대로 출력안되던 버그 수정

+++ src/screen/ScoreScreen.java
@@ -73,7 +73,7 @@ public class ScoreScreen extends Screen {
 		this.nameCharSelected = 0;
 		this.selectionCooldown = Core.getCooldown(SELECTION_TIME);
 		this.selectionCooldown.reset();
-		this.per = shipsDestroyed / bulletsShot;
+		this.per = (float)shipsDestroyed / bulletsShot;
 
 		try {
 			this.highScores = Core.getFileManager().loadHighScores();
